moved combinatoric functions into solver
