--- include/rank.h
@@ -1,7 +1,0 @@
-#include "types.h"
-
-int permutationRank(Int8 *x, int n, int k);
-
-int combinationRank(Int8 *x, int n, int k);
-
-int combinationPairRank(Int8 *x, int n);
