+++ include/solver.h
@@ -1,10 +1,5 @@
-#ifndef SOLVER_H
-#define SOLVER_H
+#include "cube.h"
 
 #define MAX_MOVES 46
 
-#include "cube.h"
-
 int solve(Cube128 c, Move *ms);
-
-#endif
