--- include/types.h
@@ -1,3 +1,0 @@
-typedef unsigned long long Int64;
-typedef unsigned char Int8;
-typedef unsigned char Int4;
